requirements.txt
openai
python-docx
PyPDF2
